Moved bot to interface side of app
