--- services/assistant/app/bot/__init__.py
@@ -1,3 +1,0 @@
-from app.bot.bot import BotFactory, BotServicesFactory, BotSettings
-
-__all__ = ["BotFactory", "BotServicesFactory", "BotSettings"]
