+++ services/assistant/app/interfaces/bot/__init__.py
@@ -1,0 +1,3 @@
+from app.interfaces.bot.bot import BotFactory, BotServicesFactory, BotSettings
+
+__all__ = ["BotFactory", "BotServicesFactory", "BotSettings"]
