+++ services/assistant/app/interfaces/bot/schemas.py
